refactor(1013): extract GatewayHaExporter to src/gateway/ (SC-001, position 23)
Fresh-extraction Cat B position 23. GatewayHaExporter (Menu #87 HA Gateway
Cluster Info) moved from MistHelper.py to src/gateway/gateway_ha_exporter.py
using the SC-001 facade pattern:

- New module: src/gateway/gateway_ha_exporter.py (6 static methods, lazy
  mh = importlib.import_module("MistHelper") for live helper access)
- Re-export alias in MistHelper.py preserves MistHelper.GatewayHaExporter
  callers (menu registration, internal references) with no rewiring
- Class body replaced by pure comment breadcrumb (NOT stub — a stub would
  rebind and clobber the alias)
- pyproject.toml coverage omit updated alphabetically

Local gate:
- black --check MistHelper.py src/gateway/gateway_ha_exporter.py: clean
- ruff check MistHelper.py src/gateway/gateway_ha_exporter.py: no issues
- py_compile: clean
- python MistHelper.py --test: 66/0/131 baseline preserved
